+++ example/lib/main_page.dart
@@ -14,7 +14,7 @@ class MainPage extends StatefulWidget {
 
 class MainPageState extends State<MainPage> {
   AppsflyerSdk _appsflyerSdk;
-  Map _oaoa;
+  Map _deepLinkData;
   Map _gcd;
 
   // called on every foreground
@@ -22,20 +22,26 @@ class MainPageState extends State<MainPage> {
   void initState() {
     super.initState();
     final AppsFlyerOptions options = AppsFlyerOptions(
-        afDevKey: DotEnv().env["DEV_KEY"],
-        appId: DotEnv().env["APP_ID"],
-        showDebug: true);
+    afDevKey: DotEnv().env["DEV_KEY"],
+    appId: DotEnv().env["APP_ID"],
+    showDebug: true);
     _appsflyerSdk = AppsflyerSdk(options);
     _appsflyerSdk.onAppOpenAttribution((res) {
       print("res: " + res.toString());
       setState(() {
-        _oaoa = res;
+        _deepLinkData = res;
       });
     });
     _appsflyerSdk.onInstallConversionData((res) {
       print("res: " + res.toString());
       setState(() {
         _gcd = res;
+      });
+    });
+    _appsflyerSdk.onDeepLinking((res){
+      print("res: " + res.toString());
+      setState(() {
+        _deepLinkData = res;
       });
     });
   }
@@ -51,23 +57,24 @@ class MainPageState extends State<MainPage> {
                   future: _appsflyerSdk.getSDKVersion(),
                   builder: (BuildContext context, AsyncSnapshot snapshot) {
                     return Text(snapshot.hasData ? snapshot.data : "");
-                  })
+                  }), 
             ],
           ),
         ),
         body: FutureBuilder<dynamic>(
             future: _appsflyerSdk.initSdk(
                 registerConversionDataCallback: true,
-                registerOnAppOpenAttributionCallback: true),
+                registerOnAppOpenAttributionCallback: true,
+                registerOnDeepLinkingCallback: true),
             builder: (BuildContext context, AsyncSnapshot snapshot) {
               if (snapshot.connectionState == ConnectionState.waiting) {
                 return Center(child: CircularProgressIndicator());
               } else {
                 if (snapshot.hasData) {
                   return HomeContainer(
-                    onAttribution: _oaoa,
                     onData: _gcd,
-                    trackEvent: logEvent,
+                    deepLinkData: _deepLinkData,
+                    logEvent: logEvent,
                   );
                 } else {
                   return Center(child: Text("Error initializing sdk"));
@@ -79,4 +86,5 @@ class MainPageState extends State<MainPage> {
   Future<bool> logEvent(String eventName, Map eventValues) {
     return _appsflyerSdk.logEvent(eventName, eventValues);
   }
+
 }